+++ src/HotelClient.php
@@ -5,18 +5,18 @@ namespace Otg\Ean;
 use GuzzleHttp\Client as HttpClient;
 use GuzzleHttp\Command\Guzzle\Description;
 use GuzzleHttp\Command\Guzzle\GuzzleClient;
+use GuzzleHttp\Command\Guzzle\Serializer;
 use Otg\Ean\RequestLocation\XmlQueryLocation;
 use Otg\Ean\Subscriber\EanError;
-use Otg\Ean\Subscriber\ResultClass;
 use Otg\Ean\Subscriber\ContentLength;
 
 /**
  * HotelClient object for executing commands against the EAN Hotel API
  *
- * @method \Otg\Ean\Result\HotelListResult getHotelList(array $arguments)
- * @method \Otg\Ean\Result\RoomAvailabilityResult getRoomAvailability(array $arguments)
- * @method \GuzzleHttp\Command\Model postReservation(array $arguments)
- * @method \GuzzleHttp\Command\Model getRoomCancellation(array $arguments)
+ * @method array getHotelList(array $arguments)
+ * @method array getRoomAvailability(array $arguments)
+ * @method array postReservation(array $arguments)
+ * @method array getRoomCancellation(array $arguments)
  * @package Otg\Ean
  */
 class HotelClient extends GuzzleClient
@@ -29,10 +29,12 @@ class HotelClient extends GuzzleClient
      */
     public static function factory($config = array())
     {
+        $description = new Description(include(__DIR__ . '/Resources/hotel-xml-v3.php'));
+
         $defaults = array(
-            'request_locations' => array(
+            'serializer' => new Serializer($description, array(
                 'xml.query' => new XmlQueryLocation('xml.query')
-            ),
+            )),
             'defaults' => array(
                 'booking_endpoint' => 'https://book.api.ean.com',
                 'general_endpoint' => 'http://api.ean.com',
@@ -47,10 +49,8 @@ class HotelClient extends GuzzleClient
         $config['defaults'] += $defaults['defaults'];
 
         $httpClient = new HttpClient();
-        $description = new Description(include(__DIR__ . '/Resources/hotel-xml-v3.php'));
 
         $client = new self($httpClient, $description, $config);
-        $client->getEmitter()->attach(new ResultClass());
         $client->getEmitter()->attach(new EanError());
         $client->getEmitter()->attach(new ContentLength());