+++ src/Subscriber/EanError.php
@@ -2,7 +2,6 @@
 
 namespace Otg\Ean\Subscriber;
 
-use GuzzleHttp\Command\Event\CommandEvents;
 use GuzzleHttp\Command\Event\ProcessEvent;
 use GuzzleHttp\Event\SubscriberInterface;
 use Otg\Ean\EanErrorException;
@@ -19,20 +18,29 @@ class EanError implements SubscriberInterface
 
     public function onProcess(ProcessEvent $event)
     {
-        $response = $event->getResponse()->xml();
+        $response = $event->getResponse();
+        if (!$response) {
+            return;
+        }
 
-        $eanError = $response->EanError ?: $response->EanWsError;
+        try {
+            $xml = $response->xml();
+        } catch (\RuntimeException $e) {
+            return;
+        }
+
+        $eanError = $xml->EanError ?: $xml->EanWsError;
 
         if ($eanError) {
 
-            $e = new EanErrorException((string) $eanError->presentationMessage);
+            $e = new EanErrorException((string) $eanError->presentationMessage, $event->getTransaction());
 
             $e->setHandling((string) $eanError->handling);
             $e->setCategory((string) $eanError->category);
             $e->setVerboseMessage((string) $eanError->verboseMessage);
             $e->setItineraryId((string) $eanError->itineraryId);
 
-            CommandEvents::emitError($event->getTransaction(), $e);
+            throw $e;
         }
     }
 }